requirements.txt
absl-py==1.4.0
astunparse==1.6.3
autopep8==1.5.5
cachetools==5.3.0

charset-normalizer==3.0.1
click==7.1.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
Flask==1.1.2
Flask-Uploads==0.2.1
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==1.1.1
matplotlib==3.6.3
mkl-fft==1.3.1

mkl-service==2.4.0
nbconvert==6.2.0
numpy<=1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.12.1
